Updated the client so that if an error is found by the backend validation during the new receipt save process, it will properly update the receipt with the returned error messages.
I also fixed one bug with the "Current Receipt" that was making it not work.

Added some logic to the "Add Payment Type" code;  If you try and add another Securiyt Deposit to a receipt that already has one, it will not let you.  The reasoning for this is that it would be impossible to match payments to deposits if there were more than one on a given receipt.

Finally, I added a small amount of formatting to the printed receipt:  If a control has a money data_type, it will format the number to look like currency.